# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/stores/{storeId}/customers/me": {
      "get": {
        "operationId": "CustomersController_getMe",
        "summary": "Get current customer profile",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer profile"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "CustomersController_updateMe",
        "summary": "Update customer profile",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile updated"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/customers/me/addresses": {
      "get": {
        "operationId": "CustomersController_listAddresses",
        "summary": "List customer addresses",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of addresses"
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "CustomersController_createAddress",
        "summary": "Create a new address",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Address created"
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/customers/me/addresses/{id}": {
      "patch": {
        "operationId": "CustomersController_updateAddress",
        "summary": "Update an address",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Address updated"
          },
          "404": {
            "description": "Address not found"
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CustomersController_deleteAddress",
        "summary": "Delete an address",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Address deleted"
          },
          "404": {
            "description": "Address not found"
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/customers/me/addresses/{id}/default": {
      "patch": {
        "operationId": "CustomersController_setDefaultAddress",
        "summary": "Set address as default",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default address updated"
          },
          "404": {
            "description": "Address not found"
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/stores/{storeId}/cart": {
      "get": {
        "operationId": "CartController_getCart",
        "summary": "Get current customer active cart",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CartController_clearCart",
        "summary": "Clear all cart items",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/cart/items": {
      "post": {
        "operationId": "CartController_addItem",
        "summary": "Add item to cart",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/cart/items/{id}": {
      "patch": {
        "operationId": "CartController_updateItem",
        "summary": "Update cart item quantity",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CartController_removeItem",
        "summary": "Remove item from cart",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/analytics/sales": {
      "get": {
        "operationId": "AnalyticsController_sales",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/analytics/products": {
      "get": {
        "operationId": "AnalyticsController_products",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/analytics/customers": {
      "get": {
        "operationId": "AnalyticsController_customers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/analytics/inventory": {
      "get": {
        "operationId": "AnalyticsController_inventory",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "ShopBuilder API",
    "description": "ShopBuilder — Shopify Clone for Local Kazakh Merchants. E-commerce platform API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste JWT access token (without \"Bearer \" prefix)"
      }
    },
...
  }
}